PDumpers: Resolved a name and PDumper number clash.
A public PDumper module was released by a third party which stole a private
allocation (and leafname) from an Acorn PDumper. Now the Acorn module is
public, there's a clash. It's easier to fix by changing the Acorn module
to be called PDumperSP25 (dumper number 25) to avoid the clash.

Resolves ticket #178

Version 1.45. Tagged as 'PDumpers-1_45'