srcbuild: Added support for a new release_autobuild phase.
This new phase will initially only be required by a very small subset of
components which do some additional functionality at the very end of a
build process in order to wrap-up the build ready for the autobuilder to
do what it wants with the results.

Admin:
  Tested on Iyonix
Author:
  srevill

Version 0.46. Tagged as 'srcbuild-0_46'