ABRelease: Modified path variable for finding local resources.
The entries on the path variable are now fully canonicalised paths in their
own right, rather than relative to the CSD - this is because creating zip
archives requires changing directory into the root of the stuff to archive
which means other resources could no longer be located on the path.

The other reason for changing this was on the off-chance it'd help with the
very odd stiffs happening in autobuilds. But the finger of blame looks more
like it is pointing at Sunfish now.

Version 0.05. Tagged as 'ABRelease-0_05'