ScrnSetup: Fix bug in "Try" button for screensavers.
If you select a screensaver from the menu and 'try' it, then decide to
cancel the plug-in, the last selected screensaver is left as the current
saver, rather than restoring what you had configured originally. A reboot
would sort it out, but that's a little OTT. Fixed.

Version 0.15. Tagged as 'ScrnSetup-0_15'