Paint: Fix crashes when dragging from the sprite filer to the filer to save sprite(s).
The use of DragASprite tended to cause an abort somewhere in the kernel's sprite handling code for some reason, changed this to instead create a temporary WIMP icon on the sprite filer in question which can than be referred to in a normal call to xfersend(). The rest of the save is then handled by the normal xfersend() handlers and the temporary icon destroyed since the WIMP takes a copy anyway.
Has the extra benefits
 * Dragging from the sprite filer now obeys the 'solid drag' flags.
 * Dragging from the sprite filer no longer crashes if dropping back on the originating filer window itself.
When dragging a selection from the filer the selection is cleared at the end of the drag, to match the behaviour of the desktop filer. When saving a selection from the 'Selection' menu, they are not deselected.
Changed to use the 'package' icon when dragging multiple sprites, or file_ff9 for singles.

Version 2.06. Tagged as 'Paint-2_06'