Paint: Colour picker help made more helpful.
Previously, the help requests would end up in the background event handler which didn't look at the window handle and would just return "This is paint" even if the handle was that of the colour picker dialogue.
Also now passes documented 0x80000000 and 0x7FFFFFFF on opening rather than zero.

Version 2.07. Tagged as 'Paint-2_07'