Env: Tweak the Filer_Boot-ing of some apps.
If these apps were already booted in a different build then the filer boot
in a new build does nothing (e.g. you'll still be using the apps in the
original build in your new build). This is a particular problem if the old
build has been deleted in the mean time...

Version 1.11. Tagged as 'Env-1_11'