DDEUtils: Fix over zealous path prefixing.
Various of OS_File, OS_Find, and OS_FSControl have subreasons which force the use of a path or path variable, which DDEUtils was not respecting. Fixes ticket #167.
Moved the ReadPrefix SWI code down with the other SWIs.
Moved the Prefix SWI code down with the other SWIs.
Moved 'strip_hats' next to 'strip_hats0'.
Added FSControl_InfoToFileType to the list of FSControl subreasons operated on.
Made the various SWI error exits use the common 'return_setv' so overall the module is smaller than in 1.66.

Version 1.69. Tagged as 'DDEUtils-1_69'