ABRelease: Added Disc and OMAP4Dev ABRelease resources
Added autobuild resources for the OMAP4Dev and Disc builds. Both of
which are untested but at least the OMAP4Dev one should be fine as
it's just a copy of the OMAP3Dev resources (with a couple of edits).

The Disc build stuff is about building a (zip compressed) Filecore
disc image of the resulting stuff. This would be good for 'burning'
straight onto a USB memory stick (e.g. with the Unix 'dd' command)
to boot machines from. WORK IN PROGRESS!

Version 0.20. Tagged as 'ABRelease-0_20'