Filer: Adjust dragging to drive icons no longer attempts to reference a deleted dir viewer.
In the case of adjust dragging from one dir viewer to another it is safe to call CloseInitiatingDir because the operation is handled entirely internally. In the case of adjust dragging to a drive icon the dir viewer must be kept until after the next wimp poll so that the selection can be retrieved and the copy operation started.
DelViews.s: made the stack frame a bit more reasonably sized
GoFiler.s: removed some unused messages from the message list
MsgsIn.s: removed the corresponding dead switch

Version 2.29. Tagged as 'Filer-2_29'