RAMFS: Declare as a floppy not a fixed disc.
In the (dusty) PRM 2-204 it states that the disc record can be used to specify the size of your media and "this is how RamFS is able to be larger than an ordinary floppy disc", which ties in with it being drive 0 since drives 0-3 are floppies.
This doesn't fix FileCore's poor handling of 1 zone fixed discs, but does make RamFS work for sizes < 964k again.
The zero init of the first zone now no longer unconditionally clears 16k, since the check on RAM area size is 8k.
Recompiled test tool 'm1time'.

Version 2.26. Tagged as 'RAMFS-2_26'