FileSwitch: *COUNT and *COPY summaries cope with > 32 bit totals.
Delete dead macro.
Ammend OS_FSControl 28 API to aloow report of 64 bit results.
On failing to get any memory at all for *COPY an attempt is made to grab an 'excessive' amount of RMA to get an error message, but this was only requesting 32MB which is now quite feasible - now asks for 2GB.

Version 2.73. Tagged as 'FileSwitch-2_73'