HForm: Reunify SCSIForm with its older sister HForm.
Hopefully reduces confusion since the differences between them is so small, and eeks a few more months of life out of the utility.
Fixed some incorrect memory clears counting words, using !, but count in 1s.
Fixed some missing internationalisation in SCSIForm
Merged messages files.
Logic is now that if only one of ADFS or SCSIFS is present on startup that will be selected by default, if both a present then the user is prompted for which to look at the default being SCSIFS.
Version number comes from messages file, and is used in the graphical banner on startup.

Version 2.60. Tagged as 'HForm-2_60'