Picker: Add ability to pass in OPTION to CMHG to allow version with no international help text to be generated.
Fix spurious null at the end of the messages file entry for "ModelListHelp", although not used when NO_INTERNATIONAL_HELP is in force the whacky lookup function in Support011 was failing because MessageTrans in RISC OS 3.10 sees that as a token delimiter and the lookup gets out of sync.

Version 0.51. Tagged as 'Picker-0_51'