DesktopBoot: Correction to Desktop files
The default run action for CMOS files was "/LoadCMOS" with "/" being a shorthand for run, which checks the run path and library (where the LoadCMOS utility is on old machines).
Since LoadCMOS is now in the boot commands module the "/" stops it working since the LoadCMOS utility is not supplied. This would still work without the "/" on a machine with no boot commands module too.

Version 0.87. Tagged as 'Boot-0_87'