Wimp: Add extra Wimp_ReadSysInfo reasons 23/24/25/26/28.
RISC OS Select has added reasons 18-27 and this change implements those that map directly to features that already exist (unimplemented ones return an error as before).
One additional reason, 28, permits the reading of the flags collected from the *WimpVisualFlags command so that applications can discover if the cheesy 3D borders are on or not.

Version 5.14. Tagged as 'Wimp-5_14'