PMonitor: Change to use CModule shared makefile.
In version 0.02 it was claimed this project compiled, but that cannot be
the case since 'url.c' was missing from the makefile and is called by
url_fetch() from the main module.
Many syntax errors fixed up (silly things like missing open braces
around functions) - the earlier log messages refer to this project being
canned but it seems like the code that was checked in didn't match what
was claimed.
Now compiles, but is likely to be of limited usefulness in the absence
of any discontinued parallel port HP inkject printers.

Version 0.03. Tagged as 'PMonitor-0_03'