TerritoryModule: Fix Territory_ReadTimeZones to behave as documented.
It was reading the current timezone CMOS and returning that in R2 then adding the difference between standard and summertime to get R3.
This meant that when in one territory it was not possible to enumerate the other territories' offsets since they all returned the current CMOS value.
Now, R2 and R3 are purely informational (looked up from the table in the module).

All terrtories on CET now call summertime the preferred "CEST".

Removed unnecessary clean rule from Makefile. Changed some one register push/pull to use macros. Changed filenames to be ".s" at the end.

Added extra switches for the other regions of the world that are wide enough to span multiple timezones. Currently there is no API to read these, so they are defined to cause an assembly error (affects USA/Canada1/Australia). Fixing that is a subject of discussion, whether to support one territory having multiple timezones or just make some extra territories?

Version 0.52. Tagged as 'Module-0_52'