TerritoryModule: A few bug fixes.
The scratch space was not large enough to hold the table for Territory_Register and overran. Buffer made bigger and assert added.
Convert time string to ordinals did not fault 29th Feb in 1900/2100/2200 due to leap year rule error.
For Japan, the SetUpEras function was never called so Japanese eras didn't work.
Duff error block being returned for SWI out of range (was using the error number instead of the address of the error block).
A couple of magic numbers swapped for header file defs.

Version 0.54. Tagged as 'Module-0_54'