TerritoryManager: Addition of support for automatic daylight saving switchover
Uses *CONFIGURE AutoDST to subsequently allow the computer to decide when to apply DST or not (the underlying CMOS bit is still set/cleared, for those applications reading it).
Messages files updated to include new syntax token.
SWI Territory_DaylightSaving and Territory_DaylightRules and Territory_ConvertTimeFormats added.
Flag definitions for all the SWIs added to exported header file.
The automatic switchover evaluates when the configure command is entered, territory is changed, or time is set (with Territory_SetTime) whether the rule is known for that year.
If no rule is known, standard time is selected, and a callback set to fire on New Year's day to try again.
If a rule is known, the next transition is calculated and a callback set.
ModHead.s
  Configure and status handlers updated for extra configuration command.
  Commoned up the OS_Byte bashing code (since the two flags bits are actually in two differenc CMOS bytes).
  *CONFIGURE <local_name_for_timezone> can now scan all timezones within a territory, not just the first.
  Magic numbers swapped for symbols where available.
  Message block and flag split into flag byte + handle word.
SWIs.s
  Daylight saving SWIs added.
  SWI Territory_ConvertTimeFormats addresses the inability to do some of the matrix of 16 combinations of local/UTC and 5 byte/ordinal conversions using the existing SWIs, and allows more conversion types to be added in the future.
TerrMgr.s
  Workspace shuffled for new functionality.
  Accept international_help switch passed in from outside.
SystemDisc
  Pixel errors in the Archimedes logo fixed.
Doc/AutoDST
Test/MultiTZTest
  Test program and SWI documentation.

Version 0.47. Tagged as 'Manager-0_47'