Podule: Fix to ReadInfo_EthernetLow/High
Was preserving R2, when in fact the read info routines use R2 to return the result value. As a side effect, had to move the podule number into R4 so it is retained when Podule_ReadInfo is used to read multiple bits at once.
Simplified ReadInfo_MEMCAddress.
Removed redundant IOTCR address and task obey scripts.

Version 1.67. Tagged as 'Podule-1_67'