FileSwitch: Fix a couple of null/bad pointer dereferences
Detail:
  s/FSControl - faff_boot_option_startup (OS_FSControl 47/48) was acting as if it was given a path variable in r3, but neither of those calls use path variables, leading to aborts in some situations due to random bits of memory being peeked at.
  s/FSUtils - When CopyWildObject performs a single object copy, it wasn't taking into account the fact that the src/dest leafname pointers could be null, potentially causing a crash if high processor vectors are in use.
Since I'm not 100% sure what the correct course of action is (in the case I saw it crash, both src and dest leafnames were null), and there aren't any bugs I know of with file copies using the wrong names, the code now uses the dest leafname if the dest leafname is null.
This sounds wrong, but matches the behaviour of the old code, since byte &00000000 is highly unlikely to contain a '*' character
Admin:
  Tested on BB-xM with high processor vectors


Version 2.77. Tagged as 'FileSwitch-2_77'