DesktopBoot: Silence SoftSCSI on RISC OS 5 with no USB
If the PreDesk directory is reached but no USB loaded (either from ROM or from a suitable podule) this silently quits.
Also no longer varies its behaviour based on Wimp$State, since by definition we're in PreDesk and the Wimp hasn't started.
Ammended the comments in the 'Hide' file, since it was describing a different location to that which SCSIFS actually looks in. It's still in the wrong place, but then with SCSIFS in ROM it doesn't work anyway.

Version 0.93. Not tagged