Config2PluginBootxxxx: Fix BootApps BootBoot BootRun to not append a new section
Previously the default 'Desktop' file in the 5.xx boot sequence had the above 3 sections demarked with "RISC OS" with a space, while this plugin was searching for one demarked with "Acorn".
Now, both use "RISCOS", so the existing settings will be correctly read in, and new settings not written in a new section when merged.

Version 0.08. Tagged as 'Bootxxxx-0_08'