Help: Append tokens for historical ROMs
Presumably when this was being actively maintained the build process extracted the help text from ROM components (such as the filer) and appended it to the application. It is assumed this has long since been lost/retired.
To produce a disc version for a given historical ROM we need the tokens for the respective ROM.
* Add tokens extracted from ROMs which !Help came with
* Recover tokens for RISC OS 3.50 from dusty CD
* Adapt makefile to allow selection

Version 2.36. Tagged as 'Help-2_36'