BuildSys: Install Config2Installer into !System.500
As the PlingSystem components result in a 26 bit APCS-R module and also outputs Config2Installer into !System.360, this avoids a SysMerge overwriting the 32 bit copy with a 26 bit one and stopping the "System" configure plugin from running when booting the disc image as RISC OS 5, while allowing pre RISC OS 5 without the updated C library to run the installer needed to update the C library.
Clear?

Version 5.81. Not tagged