ThemeDefs: Get flags for RAM loaded resources right
When the 'protect ROM sprites' option is not in use the resources are loaded with *ICONSPRITES/*TOOLSPRITES.
However, the wrong flag bits were being used to decide which to use and ended up doing the operation 3 times over for each resolution: resulting in force loading Sprites11 (the last in the DATA statements) when not in that mode, resulting in a tiny mouse pointer and lots of badly aliased icons.
Removed the search for Tools22/Tools11 (the tools encode the other resolutions inside the sprite file, not by sprite file name).
Added possibility to have EX0 EY0 paint and draw tool palettes.
Removed the manual check for ThemeSetup obey file, the boot sequence can do this for us.
Docs updated.

Version 0.03. Tagged as 'ThemeDefs-0_03'