DesktopBoot: Reorder when theme resources are loaded
Needs to be a little later on, after the desktop has been entered, in order that the pended *WIMPMODE command set in Choices:Boot.PreDesk.Configure has been acted upon (so any *ICONSPRITES commands know which is the best resolution to load).
Inlined the (optional) user setting of the theme, rather than unconditionally doing so in Choices:Boot.PreDesktop then having to go back and fix it up again later.
Tested with some different eigen factors with different sprite sets for each, to check the right one gets loaded.

Version 0.99. Tagged as 'Boot-0_99'