ThemeDefs: Different scheme to defer sprite loading
When run within the Desktop file, this app is too late to affect the banner and may also have to mess about re-iconspritesing applications from later in the boot sequence.
So, run !ThemeDefs earlier, and push the *ICON/TOOLSPRITES commands into system variables which are caught in the desktop boot file.

Version 0.04. Tagged as 'ThemeDefs-0_04'