DesktopBoot: Re-reorder the theme loading (again)
By running !ThemeDefs in the desktop file, this was too late to influence the start banner. Moved earlier again.
To solve the problem of *ICONSPRITES and *TOOLSPRITES needing to be in the final desktop mode, the desktop file looks for some system variables and does the deferred load then instead.
Tested with and without 'Protect ROM sprites' turned on.

Version 0.99. Not tagged