InetSetup: Fix to work when DHCP doesn't support -p switch
The DHCP module uses OS_ReadArgs for switch matching, but in the CMHG table it strictly checks for max-args. The recently added '-p' switch therefore should not be output by !InetSetup when the available module doesn't support it.
Load.c: parse setup files with or without -p
Save.c: use RMFind() to figure out if -p is available, and omit it when not
Also
* Bumped up the !Internet version number in the comment block written out
* Make sure DHCP is plugged in during save (and unplugged if TCP/IP is actively disabled)
* Fixed the minimum DHCP version number, !InetSetup was always outputting -w even though that wasn't supported in the version that was being RMEnsure'd

Version 0.48. Tagged as 'InetSetup-0_48'