AsmUtils: Allow No32bitCode=FALSE No26bitCode=FALSE to work
For disc targets the above condition is true, so things using AsmUtils in this situtation might be run on something with no MSR/MRS.
No32bitCode=TRUE with No26bitCode=TRUE remains invalid, but the 3 other combinations are now supported.
Binaries inspected by eye for 3 combinations.

Version 0.16. Tagged as 'AsmUtils-0_16'