ThemeDefs: Fix 'sprite not found' from CDFSFiler pre RISC OS 3.80
The swap out of the CDFSFiler sprites was previously required because entering the desktop runs Desktop_CDFSFiler which then *ICONSPRITESd its private sprites if 'protect built in icons' option was off.
Now the disc pool is loaded after *DESKTOP that left CDFSFiler with no sprites prior to RISC OS 3.80.
The logic was rendered obsolete by the 'IconChange' tool, so can be removed.

Version 0.07. Tagged as 'ThemeDefs-0_07'